requirements.txt
Django==3.1.5
djangorestframework==3.12.2
Pillow==8.1.0
urllib3==1.26.2
